+++ Test site/src/main/java/com/sloopdoge/blog/repo/ItemsRepository.java
@@ -4,5 +4,5 @@ import com.sloopdoge.blog.models.Items;
 import org.springframework.data.repository.CrudRepository;
 
 /* Interface to control data in MySql database  */
-public interface ItemsRepository extends CrudRepository<Items, Integer> {
+public interface ItemsRepository extends CrudRepository<Items, Long> {
 }
